Fixed inventory management.
Used to disable inventory keybinding by setting it to an unknown key. This caused issues for users after they went back to using the vanilla client and found their inventory still disabled. Now the old inventory keybind will be logged on startup, disabled, and finally rebinding to it's original value.