Fixed fall damage algorithim.
Fixed: https://github.com/videoerror/SurvivalTest-M0d/issues/4 Fall damage is now evaluated by checking if the player is in the air, if so, then log the player's fall position and only changes if the current player's position is greater than the previous. Then, if the player is on the ground, the fall height is evaluated as the fall position minus the current player's position on the ground. Finally, the fall damage is evaluated as the greater of zero or the fall height minus three. The fall damage is later properly subracted by the health manager.